+++ apps/src/codebridge/Editor/AiTutorChatContextButton.tsx
@@ -1,6 +1,8 @@
 import {Button} from '@code-dot-org/component-library/button';
 import React from 'react';
 
+import codebridgeI18n from '@cdo/apps/codebridge/locale';
+
 interface AiTutorChatContextButtonProps {
   saveSelectionContext: () => void;
 }
@@ -10,11 +12,8 @@ const AiTutorChatContextButton: React.FC<AiTutorChatContextButtonProps> = ({
 }) => {
   return (
     <Button
-      text="Add to AI Tutor Chat"
-      onClick={() => {
-        console.log('button');
-        saveSelectionContext();
-      }}
+      text={codebridgeI18n.addToAiTutorContext()}
+      onClick={saveSelectionContext}
       size="xs"
       type="tertiary"
       color="black"
